Refine install.sh

+++ install.sh
@@ -1,8 +1,15 @@
 #!/bin/sh
 
-git clone https://github.com/gmarik/Vundle.vim.git vim/bundle/vundle
+if [ ! -e vim/bundle/Vundle.vim ]; then
+        git clone https://github.com/gmarik/Vundle.vim.git vim/bundle/Vundle.vim
+fi
 vim +BundleInstall +qa
-mv ~/.vim ~/.vim.d.bak
-mv ~/.vimrc ~/.vimrc.bak
-ln -s vim ~/.vim
-ln -s vimrc ~/.vimrc
+if [ -e ~/.vim ]; then
+	mv ~/.vim ~/.vim.d.bak
+fi
+if [ -e ~/.vimrc ]; then
+	mv ~/.vimrc ~/.vimrc.bak
+fi
+cd
+ln -s gitrepo/vimrc/vimrc .vimrc
+ln -s gitrepo/vimrc/vim .vim
